requirements/runtime.txt
accelerate>=0.29.3
einops
fastapi
fire
mmengine-lite
numpy<2.0.0
openai
outlines<0.1.0
peft<=0.11.1
pillow
protobuf
pydantic>2.0.0
pynvml
safetensors
sentencepiece
shortuuid
tiktoken
torch<=2.4.0,>=2.0.0
torchvision<=0.18.1,>=0.15.0
transformers
triton>=2.2.0,<=3.0.0; sys_platform == "linux"
uvicorn
